+++ stress-iomix.c
@@ -1043,15 +1043,17 @@ static int stress_iomix(const stress_args_t *args)
 	ret = EXIT_SUCCESS;
 reap:
 	for (i = 0; i < SIZEOF_ARRAY(iomix_funcs); i++) {
-		if (pids[i]) {
+		if (pids[i])
 			(void)kill(pids[i], SIGALRM);
-			(void)kill(pids[i], SIGKILL);
-		}
 	}
 	for (i = 0; i < SIZEOF_ARRAY(iomix_funcs); i++) {
 		if (pids[i]) {
 			int status;
 
+			if (kill(pids[i], 0) == 0) {
+				force_killed_counter(args);
+				(void)kill(pids[i], SIGKILL);
+			}
 			(void)shim_waitpid(pids[i], &status, 0);
 		}
 	}
